src/md2html: label filename-based .md codeblocks
* src/css/monospace.css: Add a css border around pre code blocks
* src/md2html: If the pre code block has a filename, i.e.

```somelang path=123
as so
```

then 123 is added as a label above, using dotted borders,
in the style of ArchWiki. An md4c patch was developed for this.

Compared to upstream, we now use the info attribute in code blocks
for this feature.

See https://md4c.readthedocs.io/en/stable/#md4c.details.Code