content/conf/make.conf: update CPU_FLAGS_X86 via cpuid2cpuflags
Flags such as bmi2 are useful for sci-ml/ollama. This wasn't previously
registered in the package.use covering CPU flags, presumably due to a
cpuid2cpuflags update/change.

Alderlake CPUs imply that bmi2 (as shown in /proc/cpuinfo) should be
taken advantage of, but the relevant Gentoo USE flag (as triggered by
the CPU_FLAGS_X86 USE_EXPAND) also needs to be triggered.